Thesis PRISM / OECD Data Manipulation Files
These files include the Rscripts for various sorts of data manipulation and data visualisation conducted for my masters thesis at the Geneva Graduate Institute (IHEAD). Please note that the primary dataset, the PRISM dataset, is not publically available and is therefore not included in these files. To reproduce my results, apply for access to the dataset and run the various scripts. https://investmentscreening.princeton.edu/dataset